requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
azure-storage-blob==12.24.0
python-dotenv==1.0.1
pydantic==2.10.3
aiofiles==24.1.0
